pyproject.toml
[tool.poetry]
name = "messenger"
version = "0.1.0"
description = ""
authors = ["DarthZiel <Studerzhan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.115.3"}
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
alembic = "^1.13.3"
asyncpg = "^0.30.0"
fastapi-users-db-sqlalchemy = "^6.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
